#refactoring: (refactor) changed audio structure
